sprint 4
updated business logic

updated data access

+++ src/main/java/iob/logic/ActivitiesServiceJpa.java
@@ -1,49 +1,37 @@
 package iob.logic;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import iob.data.ActivityCrud;
 import iob.data.ActivityEntity;
 import iob.data.InstanceCrud;
-import iob.data.InstanceEntity;
 import iob.data.UserCrud;
-import iob.data.UserEntity;
+import iob.data.UserRole;
 import iob.logic.exceptions.BadRequestException;
+import iob.logic.exceptions.DeprecatedMethodException;
 import iob.restAPI.ActivityBoundary;
-import iob.utility.DomainWithEmail;
 import iob.utility.DomainWithId;
 import iob.utility.activity.ActivityConvertor;
 
 @Service
-public class ActivitiesServiceJpa implements ActivitiesService {
+public class ActivitiesServiceJpa extends ServiceJpa implements ExtendedActivitiesService {
 
-	private ActivityCrud activityCrud;
-	private UserCrud userCrud;
-	private InstanceCrud instanceCrud;
 	private ActivityConvertor activityConvertor;
-	private String configurableDomain;
 
 	@Autowired
 	public ActivitiesServiceJpa(InstanceCrud instanceCrud, UserCrud userCrud, ActivityCrud activityCrud,
 			ActivityConvertor activityConvertor) {
-		this.instanceCrud = instanceCrud;
-		this.userCrud = userCrud;
-		this.activityCrud = activityCrud;
+		super(userCrud, activityCrud, instanceCrud);
 		this.activityConvertor = activityConvertor;
-	}
-
-	@Value("${spring.application.name:2022b}")
-	public void setConfigurableDomain(String configurableDomain) {
-		this.configurableDomain = configurableDomain;
 	}
 
 	@Override
@@ -67,43 +55,46 @@ public class ActivitiesServiceJpa implements ActivitiesService {
 		return activityConvertor.toBoundary(activityEntity); // temp - json
 	}
 
-	@Transactional(readOnly = true)
-	private boolean checkUserIdInDB(DomainWithEmail domainWithEmail) {
-		String id = domainWithEmail.getDomain() + "_" + domainWithEmail.getEmail();
-
-		Optional<UserEntity> op = userCrud.findById(id);
-		if (op.isPresent())
-			return true;
-		return false;
+	@Override
+	public List<ActivityBoundary> getAllActivities() {
+		throw new DeprecatedMethodException(
+				"deprecated method - use getAllActivities with with paginayion instead");
 	}
 	
+	@Override
 	@Transactional(readOnly = true)
-	private boolean checkInstanceIdInDB(DomainWithId domainWithId) {
-		String id = domainWithId.getDomain() + "_" + domainWithId.getId() ;
+	public List<ActivityBoundary> getAllActivities(String userDomain, String userEmail, int size, int page) {
+		UserRole userRole = checkUserRoleInDB(userDomain, userEmail);
+		if (userRole == null)
+			throw new BadRequestException();
 		
-		Optional<InstanceEntity> op = instanceCrud.findById(id);
-		if (op.isPresent())
-			return true;
-		return false;
+		if (userRole != UserRole.ADMIN)
+			throw new BadRequestException();
+		
+		List<ActivityBoundary> allActivityBoundaries = this.activityCrud
+				.findAll(PageRequest.of(page, size, Direction.ASC, "createdTimestamp", "acticityId"))
+				.stream()
+				.map(this.activityConvertor::toBoundary).collect(Collectors.toList());
+
+		return allActivityBoundaries;
 	}
 
 	@Override
-	@Transactional(readOnly = true)
-	public List<ActivityBoundary> getAllActivities() {
-		Iterable<ActivityEntity> iter = this.activityCrud.findAll();
-
-		List<ActivityBoundary> rv = new ArrayList<>();
-		for (ActivityEntity acEntity : iter) {
-			rv.add(this.activityConvertor.toBoundary(acEntity));
-		}
-
-		return rv;
+	public void deleteAllActivities() {
+		throw new DeprecatedMethodException(
+				"deprecated method - use deleteAllActivities with user domin and user email");
 	}
-
+	
 	@Override
 	@Transactional(readOnly = false)
-	public void deleteAllActivities() {
-		activityCrud.deleteAll();
+	public void deleteAllActivities(String userDomain, String userEmail) {
+		UserRole userRole = checkUserRoleInDB(userDomain, userEmail);
+		if (userRole == null)
+			throw new BadRequestException();
+		if (userRole == UserRole.ADMIN) {
+			activityCrud.deleteAll();
+			return;
+		}
+		throw new BadRequestException();
 	}
-
 }
